+++ validator-analyse/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/hypermedia-app/hydra-validator/compare/hydra-validator-analyse@0.1.8...hydra-validator-analyse@0.2.0) (2020-03-17)
+
+**Note:** Version bump only for package hydra-validator-analyse
+
+
+
+
+
 ## [0.1.8](https://github.com/hypermedia-app/hydra-validator/compare/hydra-validator-analyse@0.1.7...hydra-validator-analyse@0.1.8) (2020-01-23)
 
 **Note:** Version bump only for package hydra-validator-analyse